Added placement of BCC and FCC structures
Made updates to class so that it could display the BCC and FCC structures and do placement of atoms for each.

Last few things in progress including adding some kind of identifiable feature to atom that moves around. Also would like to add symmetry lines in the end. Perhaps a legend to identify different atoms.

Removed unnecessary code from lines to reduce clutter. Will need to add more code comments so that I am not the only person capable of understanding this program.